requirements.txt
# Core dependencies
crewai>=0.102.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.7
pydantic>=2.5.0
textblob>=0.17.1

# Web Scraping and Parsing
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
newspaper3k>=0.2.8
lxml>=5.3.0
lxml_html_clean>=0.4.0
selenium>=4.0.0
webdriver_manager>=3.8.0

# Social Media
twikit>=0.1.0
aiohttp>=3.9.0
Mastodon.py>=1.8.1
praw>=7.7.0  # Reddit API wrapper

# News and Content
newsapi-python==0.2.6

# AI and Language Processing
openai>=1.6.1
litellm>=1.60.2
nltk>=3.8.1

# Image Processing
Pillow>=9.0.0  # For image handling

# Data Processing
numpy>=1.21.0
scipy>=1.7.0
pandas>=2.0.0
matplotlib>=3.7.0

# Testing
pytest>=7.0.0
pytest-mock>=3.10.0

# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
email-validator>=1.1.3

# Authentication
python-jose>=3.3.0

# Date/Time Handling
python-dateutil>=2.8.2

# Regex
regex>=2022.1.18

# Other
typing-extensions>=4.4.0
tqdm>=4.65.0